backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
pydantic-settings==2.1.0
gunicorn==21.2.0

# Azure Dependencies
azure-storage-blob==12.19.0
azure-search-documents==11.4.0
azure-ai-formrecognizer==3.3.2
azure-core==1.29.5

# OpenAI for Embeddings
openai>=1.0.0

# File handling
python-multipart==0.0.6
aiofiles==23.2.1